+++ xrpl/transaction/reliable_submission.py
@@ -1,6 +1,6 @@
 """High-level reliable submission methods with XRPL transactions."""
 
-from time import sleep
+from asyncio import run, sleep
 from typing import Any, Dict, cast
 
 from typing_extensions import Final
@@ -22,7 +22,7 @@ class XRPLReliableSubmissionException(XRPLException):
     pass
 
 
-def _wait_for_final_transaction_outcome(
+async def _wait_for_final_transaction_outcome(
     transaction_hash: str, client: Client
 ) -> Response:
     """
@@ -31,7 +31,7 @@ def _wait_for_final_transaction_outcome(
     validated ledger, or the transaction's lastLedgerSequence has been surpassed by the
     latest ledger sequence (meaning it will never be included in a validated ledger).
     """
-    sleep(_LEDGER_CLOSE_TIME)
+    await sleep(_LEDGER_CLOSE_TIME)
     # new persisted transaction
 
     # query transaction by hash
@@ -47,7 +47,7 @@ def _wait_for_final_transaction_outcome(
 
     if last_ledger_sequence > latest_ledger_sequence:
         # outcome is not yet final
-        return _wait_for_final_transaction_outcome(transaction_hash, client)
+        return await _wait_for_final_transaction_outcome(transaction_hash, client)
 
     raise XRPLReliableSubmissionException(
         f"The latest ledger sequence {latest_ledger_sequence} is greater than the "
@@ -55,10 +55,12 @@ def _wait_for_final_transaction_outcome(
     )
 
 
-def send_reliable_submission(transaction: Transaction, client: Client) -> Response:
+async def send_reliable_submission_async(
+    transaction: Transaction, client: Client
+) -> Response:
     """
-    Submits a transaction and verifies that it has been included in a validated ledger
-    (or has errored/will not be included for some reason).
+    Asynchronously submits a transaction and verifies that it has been included in a
+    validated ledger (or has errored/will not be included for some reason).
 
     `See Reliable Transaction Submission
     <https://xrpl.org/reliable-transaction-submission.html>`_
@@ -85,4 +87,23 @@ def send_reliable_submission(transaction: Transaction, client: Client) -> Respon
             f"Transaction failed, {result_code}: {result_message}"
         )
 
-    return _wait_for_final_transaction_outcome(transaction_hash, client)
+    return await _wait_for_final_transaction_outcome(transaction_hash, client)
+
+
+def send_reliable_submission(transaction: Transaction, client: Client) -> Response:
+    """
+    Submits a transaction and verifies that it has been included in a validated ledger
+    (or has errored/will not be included for some reason).
+
+    `See Reliable Transaction Submission
+    <https://xrpl.org/reliable-transaction-submission.html>`_
+
+    Args:
+        transaction: the signed transaction to submit to the ledger. Requires a
+            `last_ledger_sequence` param.
+        client: the network client used to submit the transaction to a rippled node.
+
+    Returns:
+        The response from a validated ledger.
+    """
+    return run(send_reliable_submission_async(transaction, client))